backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
bcrypt
pydantic
pydantic_settings
python-jose
passlib
python-multipart
websockets
pytz
paho-mqtt
python-dotenv